requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
Automat==0.7.0
autopep8==1.4.4
backcall==0.1.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
bleach==3.1.0
bs4==0.0.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.2
django-cors-headers==3.0.2
django-heroku==0.3.1
djangorestframework==3.9.4
entrypoints==0.3
fake-useragent==0.1.11
gunicorn==19.9.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.18
incremental==17.5.0
inflection==0.3.1
ipykernel==5.1.1
ipython==7.6.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.5.0
lxml==4.3.4
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==5.0.0
nbconvert==5.5.0
nbformat==4.4.0
ndg-httpsclient==0.5.1
newsapi-python==0.2.5
nltk==3.4.3
notebook==5.7.8
numpy==1.16.4
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parse==1.12.0
parsel==1.5.1
parso==0.5.0
pickleshare==0.7.5
pluggy==0.12.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psycopg2==2.8.3
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pycparser==2.19
PyDispatcher==2.0.5
pyee==6.0.0
Pygments==2.4.2
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyppeteer==0.0.25
pyquery==1.4.0
pyrsistent==0.14.11
pytest==5.0.0
pytest-runner==5.1
python-dateutil==2.8.0
pytz==2019.1
pywinpty==0.5.5
pyzmq==18.0.0
qtconsole==4.5.1
Quandl==3.4.8
queuelib==1.5.0
requests==2.21.0
requests-html==0.10.0
Scrapy==1.5.2
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
soupsieve==1.8
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.32.2
traitlets==4.3.2
Twisted==19.2.1
urllib3==1.24.3
w3lib==1.20.0
wcwidth==0.1.7
webencodings==0.5.1
websockets==7.0
whitenoise==4.1.2
widgetsnbextension==3.4.2
wincertstore==0.2
zipp==0.5.1
zope.interface==4.6.0
